In C++ build, fix "narrowing conversion inside { } is ill-formed" warning.
This also exposes an actual issue. The sound definitions may be dynamically
remapped, in which case they are not necessarily <=255 any more.

BUILD_LUNATIC.

git-svn-id: https://svn.eduke32.com/eduke32@4069 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/sounds.c
@@ -166,7 +166,7 @@ void S_MusicVolume(int32_t volume)
 void S_MenuSound(void)
 {
     static int32_t SoundNum=0;
-    uint8_t menusnds[] =
+    int32_t menusnds[] =
     {
         LASERTRIP_EXPLODE,
         DUKE_GRUNT,
